+++ src/musicscan/__init__.py
@@ -26,4 +26,4 @@
 musicscan main module
 '''
 
-__version__ = "0.2.2"
+__version__ = "0.2.3"